requirements.txt
Django==4.2.16
Pillow==10.4.0
qrcode==7.4.2
reportlab==4.2.2
python-decouple==3.8
whitenoise==6.7.0
